server/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
audioread==3.0.1
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.7
decorator==5.1.1
deepgram-sdk==2.12.0
deprecation==2.1.0
ffmpeg-python==0.2.0
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
git-filter-repo==2.45.0
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.27.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==2.1.5
marshmallow==3.22.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.1
platformdirs==4.3.2
pooch==1.8.2
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.4
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.0.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
srt==3.5.3
sympy==1.13.2
threadpoolctl==3.5.0
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websockets==13.0.1
yarl==1.11.1
yt-dlp==2024.8.6
